% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corMatrix.r
\name{corMatrix}
\alias{corMatrix}
\title{Correlation Matrix and tree}
\usage{
corMatrix(filename, output = paste0(strsplit(filename, ".", fixed =
  TRUE)[[1]][1], "_Correlation.pdf"), na = F, landscape = F,
  dendrogram = T, heatmap = T, compound = NULL, metabolite = NULL,
  sampleStart = 3, corDigit = 3, testType = "spearman",
  plotWidth = 0.5, plotHeigth = 0.2, textSize = 0.3,
  upperLimit = 0.6, lowerLimit = -0.6)
}
\arguments{
\item{filename}{name of the datafile
    Must contain:
\itemize{
\item a column "Compound";
\item a column "Metabolite";
\item and all the columns sample from <sampleStart> to the end of the file;
}
the rest doesn't matter and the names are optional, as long as the column position is
entered.;}

\item{output}{default <dataset filename>.pdf, name of the pdf file;}

\item{na}{default FALSE, FALSE removes the untargeted meatabolite;}

\item{landscape}{default FALSE, orientation of the pdf file;}

\item{dendrogram}{default TRUE, presence of the correlation dendrogram;}

\item{heatmap}{default TRUE, presence of the correlation matrix heatmap;}

\item{compound}{default NULL, position of the compound column if named otherwise;}

\item{metabolite}{default NULL, position of the metabolite column if named otherwise;}

\item{sampleStart}{default 3, 1st column of the actual data.}

\item{corDigit}{default 3,number of digit for the correlation coefficient;}

\item{testType}{default spearman, type of correlation test. Can be "spearman"or "pearson";}

\item{plotWidth}{default 0.5, width of the top space above the heatmap;}

\item{plotHeigth}{default 0.2, width of the side space beside the heatmap;}

\item{textSize}{default 2, size of the text in the heatmap tiles and the dendrogram;}

\item{upperLimit}{default 0.6, positive correlation limit to output the metabolite pair;}

\item{lowerLimit}{default -0.6, negative correlation limit to output the metabolite pair.}
}
\value{
the rcorr objet, with:
\itemize{
\item \code{r} The correlation matrix
\item \code{P} The P-value matrix
}
}
\description{
Creates the correlation matrix and tree for a metabolomic dataset.
Produces the following files:
\itemize{
\item <filename>.pdf, containing the correlation matrix
\item <filename>_pairs.txt, containing the correlation pairs above <upperLimit> or less than <lowerLimit>;
}
}
\details{
The significance of the correlation is noted with a "*" in the tile.
Conditions for the Pearson's Correlation Test:
\itemize{
\item Independent samples;
\item normal distribution of the data; *
}
*Since this condition is not true for most metabolite,
       the test is set as a Spearman's rank Correlation test by default.*
}
\examples{
for a dataset with the following header ; Compound, m/z, Metabolite, RT, Sample #1, ...
corMatrix("dummySet.tsv", sampleStart=5, testType="spearman")

for a dataset with the following header ; compound, m/z, metabolite, RT, Sample #1, ...
corMatrix("dummySet.tsv", compound=1, metabolite=3, sampleStart=5, testType="spearman")

}
\seealso{
\itemize{
\item \code{reshape2} package\url{https://www.rdocumentation.org/packages/reshape2}
\item \code{Hmisc} package \url{https://www.rdocumentation.org/packages/Hmisc}
\item \code{gplot} package \url{https://www.rdocumentation.org/packages/gplot}
}
}
