% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr_computation.R
\name{egfr_calculation}
\alias{egfr_calculation}
\title{A function to compute estimated glomular filtration rate eGFR}
\usage{
egfr_calculation(df, sex_colname, creatine_colname, age_colname)
}
\arguments{
\item{df}{dataframe with individuals in rows and phenotypes (blood measures, age, sex, ect...) in columns}

\item{sex_colname}{column name for gender value}

\item{creatine_colname}{column name for creatine value}

\item{age_colname}{age column name for age value}
}
\value{
dataframe with eGFR values as scores
}
\description{
Calculate eGFR
}
\examples{
calculate eGFR from dataset with phenotypes
new_df <- egfr_calculation(df,sex_colname="Sex",
          creatine_colname="Creatinine",
          age_colname="Age")
}
