% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputation}
\alias{imputation}
\title{Metabolomic Dataset Imputation}
\usage{
imputation(file, k = 2, method = "knn", npcs = 3, sigma = 0.1,
  nTree = 30, na.string = "NA", transformation = "None",
  compound = NULL, metabolite = NULL, sampleStart = 3)
}
\arguments{
\item{file}{file containing the dataset to impute;
\itemize{
\item a column "Compound";
\item a column "Metabolite";
\item and all the columns sample from <sampleStart> to the end of the file;
}
      the rest doesn't matter and the names are optional, as long as the column position is entered.}

\item{k}{default 2, the k used for the knn imputation;}

\item{method}{default "knn", the chosen method for replacing the missing values. Can be "knn", "RF",
"QRILC", "SVD", "mean", "median", "HM" or "0". See Details.}

\item{npcs}{default 3, npcs for SVD method;}

\item{sigma}{default 0.1, tune sigma parameter for QRILC method;}

\item{nTree}{default 30, number of tree for the RF method;}

\item{na.string}{default "NA", string to consider as NA in the dataset;}

\item{transformation}{default "None", can be "scale" or "log";}

\item{compound}{default NULL, position of the compound column if named otherwise;}

\item{metabolite}{default NULL, position of the metabolite column if named otherwise;}

\item{sampleStart}{default 3, 1st column of the actual data;}
}
\value{
df, the imputed dataset as a dataframe.
}
\description{
Impute the given dataset with different method options.
Produces <filename>_imputed.txt, containing the imputed dataset;
See Details for available Imputation Methods
}
\details{
Available imputation methods:
\itemize{
\item "knn": From the \code{impute} package, use the k nearest neighboors to impute the values;
\item "RF": From the \code{missForest} package, use RandomForest algorithm to impute the values;
\item "QRILC": From the \code{imputeLCMD} package, use Quantile regression to impute the values;
\item "SVD": From the \code{pcaMethods} package, use SVDimpute algorithm as proposed by Troyanskaya et al, 2001. to impute the values;
\item "mean","median", ""median", "0", "HM": simple value replacement, either by the mean, median, 0 of Half minimum of the row;
}
}
\examples{
for a dataset with the following header ; Compound, m/z, Metabolite, RT, Sample #1, ...
imputation("dummySet.tsv", method="knn", transformation="log", sampleStart=5)

for a dataset with the following header ; compound, m/z, metabolite, RT, Sample #1, ...
imputation("dummySet.tsv", method="knn", transformation="log", metabolite = 3, compound = 1, sampleStart=5)


}
\seealso{
\itemize{
\item \code{impute} package\url{https://www.rdocumentation.org/packages/impute}
\item \code{missForest} package \url{https://www.rdocumentation.org/packages/missForest}
\item \code{imputeLCMD} package \url{https://www.rdocumentation.org/packages/imputeLCMD}
\item \code{pcaMethods} package \url{https://www.rdocumentation.org/packages/pcaMethods}
}
}
