% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputationTest}
\alias{imputationTest}
\title{Test for the various methods}
\usage{
imputationTest(input, output = NULL, k = 2, method = "knn",
  npcs = 3, sigma = 0.1, nbTest = 10, nTree = 30,
  na.string = "NA", missing = 0.05, missingType = "MCAR",
  transformation = "None", sampleStart = 3)
}
\arguments{
\item{input}{file containing the test dataset; Should contain:
\itemize{
\item all the columns sample from <sampleStart> to the end of the file;
}
*MUST BE COMPLETE (no NA values) for accurate results}

\item{output}{default NULL, name of the test results file if not NULL}

\item{k}{default 2, the k used for the knn imputation;}

\item{method}{default "knn", the chosen method for replacing the missing values. Can be "knn", "RF",
"QRILC", "SVD", "mean", "median", "HM" or "0". See Details.}

\item{npcs}{default 3, npcs for SVD method;}

\item{sigma}{default 0.1, tune sigma parameter for QRILC method;}

\item{nTree}{default 30, number of tree for the RF method;}

\item{na.string}{default "NA", string to consider as NA in the dataset;}

\item{missing}{default 0.05, proportion of missing values;}

\item{missingType, }{default "MCAR" (missing completely at random), can be "MNAR" (not at ramdom), will target the values under the median;}

\item{transformation}{default "None", can be "scale" or "log";}

\item{sampleStart}{default 3, 1st column of the actual data;}

\item{nTest, }{default 10, number of test to loop;}

\item{compound}{default NULL, position of the compound column if named otherwise;}

\item{metabolite}{default NULL, position of the metabolite column if named otherwise;}
}
\value{
resDf, the result NRMSE dataframe.
}
\description{
Test imputation accuracy for the given dataset with different method options. Compute the NRMSE values for the desired number of test.
If asked, produces an output <filename>_Accuracy.txt with the columns
\itemize{
\item "Method";
\item "missing_proportion";
\item "transformation";
\item  and	"NRMSE"
}
}
\details{
Will compute de NRMSE (Normalized Root Mean Squared Error) for an imputation test.
Available methods:
\itemize{
\item "knn": From the \code{impute} package, use the k nearest neighboors to impute the values;
\item "RF": From the \code{missForest} package, use RandomForest algorithm to impute the values;
\item "QRILC": From the \code{imputeLCMD} package, use Quantile regression to impute the values;
\item "SVD": From the \code{pcaMethods} package, use SVDimpute algorithm as proposed by Troyanskaya et al, 2001. to impute the values;
\item "mean","median", ""median", "0", "HM": simple value replacement, either by the mean, median, 0 of Half minimum of the row;
}
}
\examples{
for a dataset with the following header ; Compound, m/z, Metabolite, RT, Sample #1, ...
imputationTest("dummySet.tsv", method="knn", transformation="log", sampleStart=5)

for a dataset with the following header ; compound, m/z, metabolite, RT, Sample #1, ...
imputationTest("dummySet.tsv", method="knn", transformation="log", metabolite = 3, compound = 1, sampleStart=5)


}
\seealso{
\itemize{
\item \code{impute} package\url{https://www.rdocumentation.org/packages/impute}
\item \code{missForest} package \url{https://www.rdocumentation.org/packages/missForest}
\item \code{imputeLCMD} package \url{https://www.rdocumentation.org/packages/imputeLCMD}
\item \code{pcaMethods} package \url{https://www.rdocumentation.org/packages/pcaMethods}
}
}
