% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_pval.R
\name{linearRegcov_p}
\alias{linearRegcov_p}
\title{A functions to extract p-value from linear regressions}
\usage{
linearRegcov_p(y, ...)
}
\arguments{
\item{y}{numerical vector, depentent variable}

\item{x}{numerical vector, explanatory variable}
}
\value{
p-value from linear regressions
}
\description{
Extract p-value from linear regression
}
\examples{
#linear regression pvalue
pval <- linearRegcov_p(y = mtcars$mpg,x=mtcars$cyl)
#linear regression pvalue from data frame
observed_pval <- sapply(x=df, linearRegcov_p, y=trait, data=covariable)
}
