% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_stderr.R
\name{linearRegcov_stderr}
\alias{linearRegcov_stderr}
\title{A function to extract standard error from linear regressions}
\usage{
linearRegcov_stderr(y, ...)
}
\arguments{
\item{y}{numerical vector, depentent variable}

\item{x}{numerical vector, explanatory variable}
}
\value{
standard error from regressions
}
\description{
Extract standard error from linear regression
#'
}
\examples{
#linear regression standard error
standard_error <- linearRegcov_stderr(y = mtcars$mpg,x=mtcars$cyl)
#linear regression pvalue to a data frame
standard_error <- sapply(x=df, linearRegcov_stderr, y=traitdf$y, data=covariable)
}
