% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression_confidence_interval_lower.R
\name{logisticRegcov_confint_lower}
\alias{logisticRegcov_confint_lower}
\title{A functions to extract confidence interval lower bound from logistic regressions}
\usage{
logisticRegcov_confint_lower(y, ...)
}
\arguments{
\item{y}{numerical vector, depentent variable}

\item{x}{numerical vector, explanatory variable}
}
\value{
confidence interval lower regressions
}
\description{
Extract CI-L from logistic regression
}
\examples{
#logistics regression confidence interval lower from logistic regression
confidence_interval_lower <- logisticRegcov_confint_lower(y = mtcars$am ,x=mtcars$mpg)
#linear regression confidence interval lower bounds
confidence_interval_lower <- sapply(x=df, logisticRegcov_confint_lower, y=df$y, data=covariable)
}
