% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression_confidence_interval_upper.R
\name{logisticRegcov_confint_upper}
\alias{logisticRegcov_confint_upper}
\title{A functions to extract confidence interval upper bound from logistic regressions}
\usage{
logisticRegcov_confint_upper(y, ...)
}
\arguments{
\item{y}{numerical vector, depentent variable}

\item{x}{numerical vector, explanatory variable}
}
\value{
confidence interval upper regressions
}
\description{
Extract CI-U from logistic regression
}
\examples{
#logistics regression confidence interval upper from logistic regression
confidence_interval_upper <- logisticRegcov_confint_upper(y = mtcars$am ,x=mtcars$mpg)
#linear regression confidence interval upper bounds
confidence_interval_upper <- sapply(x=df, logisticRegcov_confint_upper, y=df$y, data=covariable)
}
