% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression_odds_ratio.R
\name{logisticRegcov_effect}
\alias{logisticRegcov_effect}
\title{A function to extract odds ratio from logistic regressions}
\usage{
logisticRegcov_effect(y, ...)
}
\arguments{
\item{y}{numerical vector, depentent variable}

\item{x}{numerical vector, explanatory variable}
}
\value{
odds ratio from regressions
}
\description{
Extract odds ratio from logistic regression
}
\examples{
#logistics regression odds ratio from logistic regression
confidence_interval_upper <- logisticRegcov_effect(y = mtcars$am ,x=mtcars$mpg)
#linear regression odds ratio
odds_ratio <- sapply(x=df, logisticRegcov_effect, y=df$y, data=covariable)
}
