% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression_pval.R
\name{logisticRegcov_p}
\alias{logisticRegcov_p}
\title{A function to extract p-value from logistic regressions}
\usage{
logisticRegcov_p(y, ...)
}
\arguments{
\item{y}{numerical vector, depentent variable}

\item{x}{numerical vector, explanatory variable}
}
\value{
p-value from regressions
}
\description{
Extract p-value from logistic regression
}
\examples{
#logistics regression p-value from logistic regression
pval <- logisticRegcov_p(y = mtcars$am ,x=mtcars$mpg)
#linear regression p-value
pval <- sapply(x=df, logisticRegcov_p, y=df$y, data=covariable)
}
