% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression_sample_size.R
\name{logisticRegcov_samplesize}
\alias{logisticRegcov_samplesize}
\title{A function to extract sample size from logistic regression}
\usage{
logisticRegcov_samplesize(y, ...)
}
\arguments{
\item{y}{numerical vector, depentent variable}

\item{x}{numerical vector, explanatory variable}
}
\value{
sample size from logistic regressions
}
\description{
Extract sample size logistic regression
}
\examples{
#logistic regression pvalue
pval <- linearRegcov_p(y = mtcars$mpg,x=mtcars$cyl)
#logistic regression sample size to a data frame
sample_size <- sapply(x=df, logisticRegcov_samplesize, y=df$y, data=covariable)
}
