% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{NRMSE}
\alias{NRMSE}
\title{NRMSE computing}
\usage{
NRMSE(dat.impute, dat.true)
}
\arguments{
\item{dat.impute}{imputed dataframe}

\item{dat.true}{original dataframe}
}
\value{
NRMSE value
}
\description{
Compute NRMSE for accuracy mesure. Compare tow matrices, one being imputed from the original one with holes added.
}
\details{
NRMSE is sqrt((sum((dat.impute-dat.true)^2) / totalNbValues ) / variance(dat.true))
}
\examples{
NRMSE(ImputedDataset, OriginalCompleteDataset)

}
