% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.r
\name{QCcustomization}
\alias{QCcustomization}
\title{Customization of QC output}
\usage{
QCcustomization(data, COMPOUND = TRUE, COHORTE = TRUE,
  METABOLITE = TRUE, NBNA = TRUE, SD = TRUE, MU = TRUE,
  CV = TRUE, REMOVE = TRUE)
}
\arguments{
\item{data}{the input QC dataframe;}

\item{COMPOUND}{default=TRUE, output the column named "compound";}

\item{NBNA}{default=TRUE, output the column named "nbna";}

\item{SD}{default=TRUE, output the column named "sd";}

\item{MU}{default=TRUE, output the column named "mu";}

\item{CV}{default=TRUE, output the column named "CV";}

\item{REMOVE}{default=TRUE, output the column named "remove".}
}
\value{
subdata the desired dataframe,
}
\description{
Enable the user to choose which parameter to output from a dataframe containing the columns compound, metabolite, nbna,	sd,	mu,	CV,	and remove.
}
\examples{
QCcustomization(QCresult$QC, REMOVE=FALSE)

}
