% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.r
\name{qualityControl}
\alias{qualityControl}
\title{Data Quality Control}
\usage{
qualityControl(data, output = NULL, cohort = "", missing = 0.2,
  compound = NULL, metabolite = NULL, sampleStart = 6)
}
\arguments{
\item{data}{name of the initial metabolomic dataset file.
Must contain:
  -a column "Compound";
  -a column "Metabolite";
  -and all the columns sample from <sampleStart> to the end of the file;
  the rest doesn't matter and the names "Compound" and "Metabolite" are optional, as long as the column position is specified in params. ;}

\item{output}{default NULL, name of the output cleaned dataset to produce if not NULL;}

\item{cohort}{default "sample", name of the patient cohort;}

\item{compound}{default NULL, position of the compound column if named otherwise;}

\item{metabolite}{default NULL, position of the metabolite column if named otherwise;}

\item{sampleStart}{default 6, the index of the 1st sample column.}
}
\value{
a list, containing the cleaned dataset and the QC dataframe
\itemize{
\item dataset cleaned dataset
\item QC the QC dataframe containing  the columns compound, metabolite, nbna,	sd,	mu,	CV,	 and remove.

}
}
\description{
Quality control for metabolomic data. Will remove metabolites with more than a determined thershold of missing values for all patient.
Produces the foloowing files:
  -"QC_ <data> .txt": Summary of the QC;
  -"REMOVED_QC_<data>.txt": Summary of the removed metabolite;
  - output, cleaned dataset file(optional);
  -"REMOVED_<output>.txt", the removed set of metabolite (optional)
}
\details{
Can be used on any kind of file containing rows with x (gene, metabolite, etc.) and column with y (Sample, patients, etc.).
}
\examples{
for a dataset with the following header ; Compound, m/z, Metabolite, RT, Sample #1, ...
qualityControl("dummySet.txt", sampleStart = 5)

for a dataset with the following header ; compound, m/z, metabolite, RT, Sample #1, ...
qualityControl("dummySet.txt", sampleStart = 5, compound = 1, metabolite = 3)


}
