% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness_drop_samples.R
\name{robustness_drop_samples}
\alias{robustness_drop_samples}
\title{A function to visualy inspect module robustness}
\usage{
robustness_drop_samples(metabolites_data, module_colors, power1, main1,
  filename)
}
\arguments{
\item{metabolites_data}{metabolite dataframe}

\item{module_colors}{list of module colors generated}

\item{power1}{power of scale free topology}

\item{main1}{name of the plots}

\item{filename}{filename of the output plot}
}
\value{
plot
}
\description{
Testing robustness of modules by dropping half of samples
}
\examples{
set.seed(1)
pdf(file = "Robustness_dropping_samples.pdf", width = 12, height = 9)
par(mfrow=c(8,2), mar=c(2,2,2,1))
for (i in 1:8) {
 subsetindivs=sample(1:no.indivs,344,replace=F)
 robustness_drop_samples(metabolites_data = allsamples_norm_cl_trimmed[subsetindivs,],
                      module_colors = moduleColors,power1=6 ,main1="SCD Metabolites Network",
                      filename=paste("metabolites_clust_",i,".txt",sep=""))
}
dev.off()
}
