% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{run_combat}
\alias{run_combat}
\title{A function to apply combat to your dataframe konwing the batch}
\usage{
run_combat(df, batch_df, batch_colnumb)
}
\arguments{
\item{df}{dataframe with individuals in rows and metabolites in columns}

\item{batch_df}{dataframe with information on batches (which indivudual belong to which batch)}

\item{batch_colname}{colname for actual batch}
}
\value{
dataframe to which ComBat was applied
}
\description{
Apply batch effect correction using function ComBat
}
\examples{
removing batch effect from dataset using combat
new_df <- batch_combat(allsamples, batch_data, individual_batches)
}
