% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplots_metabolite_connectivity_strength.R
\name{scatterplots_metabo}
\alias{scatterplots_metabo}
\title{A function plot module connection strength}
\usage{
scatterplots_metabo(module_color, trait, metabolite_significance,
  module_membership_cor, alt_color)
}
\arguments{
\item{module_color}{module color}

\item{trait}{name of trait}

\item{alt_color}{module color different from the original module name}

\item{dataset}{dataframe with module membership and metabolite significance}
}
\value{
plot
}
\description{
Plotting Module membership (x-axis) and metabolite significance (y-axis)
}
\examples{
scatterplots_metabo(module_color="darkorange",trait="RBC.beta",
dataset=metabo_bloodtrait_cor_final,alt_color="")
scatterplots_metabo(module_color="lightcyan",trait="eGFR.beta",
dataset="metabo_bloodtrait_cor_final",module_membership_cor="module_membership_cor",alt_color="blue")
}
