% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinPlot.R
\name{violinPlotImp}
\alias{violinPlotImp}
\title{Comparative Violin Plot Constructor}
\usage{
violinPlotImp(input, inputImputed = paste0(strsplit(input, ".", fixed =
  TRUE)[[1]][1], "_imputed.txt"), output = paste0(strsplit(input, ".",
  fixed = TRUE)[[1]][1], "_Imputed.pdf"), na = F, numberByGraph = 1,
  colNumberByPage = 5, rowNumberByPage = 2, landscape = F,
  compound = NULL, metabolite = NULL, sampleStart = 3,
  compoundImp = NULL, metaboliteImp = NULL, sampleStartImp = 3,
  onlyImputed = TRUE)
}
\arguments{
\item{input}{name of the cleaned datafile;
Must contain:
 -a column named "Compound";
 -a column named "Metabolite";
 -and all the columns sample from <sampleStart> to the end of the file;
  the rest doesn't matter and the names are optional, as long as the column position is entered..}

\item{inputImputed}{name of the imputed datafile;
    Must contain:
\itemize{
\item a column "Compound";
\item a column "Metabolite";
\item and all the columns sample from <sampleStart> to the end of the file;
}
the rest doesn't matter and the names are optional, as long as the column position is entered.}

\item{output}{default <input>.pdf, name of the final pdf file;}

\item{na}{default FALSE, FALSE removes the untargeted (NA) metabolite from the plots;}

\item{numberByGraph}{default 1, number of metabolite violin plots by graph window;}

\item{colNumberByPage}{default 5, number of column in the output pdf file;}

\item{rowNumberByPage}{default 2, number of row in the output pdf file;}

\item{landscape}{default FALSE, orientation of the pdf file;}

\item{compound}{default NULL, position of the compound column if named otherwise;}

\item{metabolite}{default NULL, position of the metabolite column if named otherwise;}

\item{sampleStart}{default 3, 1st column of the compounds intensity.}

\item{compoundImp}{default NULL, position of the compound column for Imputed dataset if named otherwise;}

\item{metaboliteImp}{default NULL, position of the metabolite column for Imputed dataset if named otherwise;}

\item{sampleStartImp}{default 3, 1st column of the compounds intensity for Imputed dataset.}

\item{onlyImputed}{default TRUE, TRUE will display only plots for metabolites with imputed values. (nb_imputed > 0)}
}
\value{
None
}
\description{
Construct Violin plots from imputed dataset and cleaned by QC.r.
Produces a pdf File containing the plots.
}
\details{
it is possible to put any 2 corresponding files, as long as they have the same row and column names. (The imputation part is not necessary, but it is the principal input type of this function. The legend and labels does contain the term "Imputed" by default.)
}
\examples{
for 2 dataset with the following headers ; Compound, Metabolite, Sample #1, ...
violinPlotQC<- function("dummySet.txt")

for an imputed dataset with the following header ; compound, metabolite,  Sample #1, ...
violinPlotQC<- function("dummySet.txt", compoundImp = 1, metaboliteImp = 2)

}
\seealso{
\itemize{
\item \code{ggplot2} package\url{https://www.rdocumentation.org/packages/ggplot2}
\item \code{tidyr} package \url{https://www.rdocumentation.org/packages/tidyr}
\item \code{gridExtra} package \url{https://www.rdocumentation.org/packages/gridExtra}
\item \code{ggpubr} package \url{https://www.rdocumentation.org/packages/ggpubr}
\item \code{scales} package \url{https://www.rdocumentation.org/packages/scales}
}
}
