% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinPlot.R
\name{violinPlotQC}
\alias{violinPlotQC}
\title{Simple Violin Plot Constructor}
\usage{
violinPlotQC(input, output = paste0(strsplit(input, ".", fixed =
  TRUE)[[1]][1], ".pdf"), na = F, numberByGraph = 1,
  colNumberByPage = 5, rowNumberByPage = 2, landscape = F,
  compound = NULL, metabolite = NULL, sampleStart = 3)
}
\arguments{
\item{input}{name of the cleaned datafile;
    Must contain:
\itemize{
\item a column "Compound";
\item a column "Metabolite";
\item and all the columns sample from <sampleStart> to the end of the file;
}
the rest doesn't matter and the names are optional, as long as the column position is entered.}

\item{output}{default <input>.pdf, name of the final pdf file;}

\item{na}{default FALSE, FALSE removes the untargeted (NA) metabolite from the plots;}

\item{numberByGraph}{default 1, number of metabolite violin plots by graph window;}

\item{colNumberByPage}{default 5, number of column in the output pdf file;}

\item{rowNumberByPage}{default 2, number of row in the output pdf file;}

\item{landscape}{default FALSE, orientation of the pdf file;}

\item{compound}{default NULL, position of the compound column if named otherwise;}

\item{metabolite}{default NULL, position of the metabolite column if named otherwise;}

\item{sampleStart}{default 3, 1st column of the compounds intensity.}
}
\value{
None
}
\description{
Construct Violin plots for a metabolomic dataset cleaned by QC.r (not required).
Produces a pdf File containing the plots.
}
\details{
Can be used on any kind of file containing rows with x (gene, metabolite, etc.) and column with y (Sample, patients, etc.).
}
\examples{
for a dataset with the following header ; Compound, m/z, Metabolite, RT, Sample #1, ...
violinPlotQC<- function("dummySet.txt", sampleStart = 5)

for a dataset with the following header ; compound, m/z, metabolite, RT, Sample #1, ...
violinPlotQC<- function("dummySet.txt", compound = 1, metabolite = 3, sampleStart = 5)

}
\seealso{
\itemize{
\item \code{ggplot2} package\url{https://www.rdocumentation.org/packages/ggplot2}
\item \code{tidyr} package \url{https://www.rdocumentation.org/packages/tidyr}
\item \code{gridExtra} package \url{https://www.rdocumentation.org/packages/gridExtra}
\item \code{ggpubr} package \url{https://www.rdocumentation.org/packages/ggpubr}
\item \code{scales} package \url{https://www.rdocumentation.org/packages/scales}
}
}
